+++ src/migration/ember-engine/steps/create-file-path-maps.ts
@@ -1,8 +1,8 @@
 import type { FilePathMap } from '@codemod-utils/files';
 
 import type { Options } from '../../../types/index.js';
-import { migrationStrategyForAddonFolder } from './create-file-path-maps/addon/index.js';
-import { migrationStrategyForTestsFolder } from './create-file-path-maps/tests/index.js';
+import { mapAddonFolder } from './create-file-path-maps/addon/index.js';
+import { mapTestsFolder } from './create-file-path-maps/tests/index.js';
 
 type MigrationStrategies = {
   addon: FilePathMap;
@@ -10,11 +10,8 @@ type MigrationStrategies = {
 };
 
 export function createFilePathMaps(options: Options): MigrationStrategies {
-  const filePathMapAddon = migrationStrategyForAddonFolder(options);
-  const filePathMapTests = migrationStrategyForTestsFolder(options);
-
   return {
-    addon: filePathMapAddon,
-    tests: filePathMapTests,
+    addon: mapAddonFolder(options),
+    tests: mapTestsFolder(options),
   };
 }
