+++ src/migration/ember-engine/steps/create-file-path-maps/addon/map-route-routes.ts
@@ -5,23 +5,21 @@ import type {
   Options,
 } from '../../../../../types/index.js';
 
-export function migrationStrategyForComponentTemplates(
-  options: Options,
-): FilePathMapEntries {
+export function mapRouteRoutes(options: Options): FilePathMapEntries {
   const { projectRoot } = options;
 
-  const filePaths = findFiles('addon/components/**/template.hbs', {
+  const filePaths = findFiles('addon/**/route.{js,ts}', {
     projectRoot,
   });
 
   return filePaths.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
-        directory: 'addon/components',
-        file: 'template',
+        directory: 'addon',
+        file: 'route',
       },
       replace: (key: string) => {
-        return `addon/components/${key}`;
+        return `addon/routes/${key}`;
       },
     });
