+++ src/migration/ember-engine/steps/create-file-path-maps/tests/index.ts
@@ -1,12 +1,12 @@
 import type { FilePathMap, Options } from '../../../../../types/index.js';
-import { migrationStrategyForComponents } from './components.js';
-import { migrationStrategyForRouteControllers } from './route-controllers.js';
-import { migrationStrategyForRouteRoutes } from './route-routes.js';
+import { mapComponents } from './map-components.js';
+import { mapRouteControllers } from './map-route-controllers.js';
+import { mapRouteRoutes } from './map-route-routes.js';
 
-export function migrationStrategyForTestsFolder(options: Options): FilePathMap {
+export function mapTestsFolder(options: Options): FilePathMap {
   return new Map([
-    ...migrationStrategyForComponents(options),
-    ...migrationStrategyForRouteControllers(options),
-    ...migrationStrategyForRouteRoutes(options),
+    ...mapComponents(options),
+    ...mapRouteControllers(options),
+    ...mapRouteRoutes(options),
   ]);
 }
