+++ src/migration/ember-engine/steps/create-file-path-maps/tests/map-route-controllers.ts
@@ -5,26 +5,27 @@ import type {
   Options,
 } from '../../../../../types/index.js';
 
-export function migrationStrategyForRouteRoutes(
-  options: Options,
-): FilePathMapEntries {
+export function mapRouteControllers(options: Options): FilePathMapEntries {
   const { projectRoot } = options;
 
   /*
     Case 1: Didn't pass the --pod flag, but configured { usePods: true } in .ember-cli
   */
-  const filePaths1 = findFiles('tests/unit/!(routes)/**/route-test.{js,ts}', {
-    projectRoot,
-  });
+  const filePaths1 = findFiles(
+    'tests/unit/!(controllers)/**/controller-test.{js,ts}',
+    {
+      projectRoot,
+    },
+  );
 
   const filePathMap1 = filePaths1.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
         directory: 'tests/unit',
-        file: 'route-test',
+        file: 'controller-test',
       },
       replace: (key: string) => {
-        return `tests/unit/routes/${key}-test`;
+        return `tests/unit/controllers/${key}-test`;
       },
     });
 
@@ -34,18 +35,21 @@ export function migrationStrategyForRouteRoutes(
   /*
     Case 2: Passed the --pod flag to Ember CLI
   */
-  const filePaths2 = findFiles('tests/unit/routes/**/route-test.{js,ts}', {
-    projectRoot,
-  });
+  const filePaths2 = findFiles(
+    'tests/unit/controllers/**/controller-test.{js,ts}',
+    {
+      projectRoot,
+    },
+  );
 
   const filePathMap2 = filePaths2.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
-        directory: 'tests/unit/routes',
-        file: 'route-test',
+        directory: 'tests/unit/controllers',
+        file: 'controller-test',
       },
       replace: (key: string) => {
-        return `tests/unit/routes/${key}-test`;
+        return `tests/unit/controllers/${key}-test`;
       },
     });
