+++ src/migration/ember-engine/steps/create-file-path-maps/tests/map-route-routes.ts
@@ -5,29 +5,24 @@ import type {
   Options,
 } from '../../../../../types/index.js';
 
-export function migrationStrategyForRouteControllers(
-  options: Options,
-): FilePathMapEntries {
+export function mapRouteRoutes(options: Options): FilePathMapEntries {
   const { projectRoot } = options;
 
   /*
     Case 1: Didn't pass the --pod flag, but configured { usePods: true } in .ember-cli
   */
-  const filePaths1 = findFiles(
-    'tests/unit/!(controllers)/**/controller-test.{js,ts}',
-    {
-      projectRoot,
-    },
-  );
+  const filePaths1 = findFiles('tests/unit/!(routes)/**/route-test.{js,ts}', {
+    projectRoot,
+  });
 
   const filePathMap1 = filePaths1.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
         directory: 'tests/unit',
-        file: 'controller-test',
+        file: 'route-test',
       },
       replace: (key: string) => {
-        return `tests/unit/controllers/${key}-test`;
+        return `tests/unit/routes/${key}-test`;
       },
     });
 
@@ -37,21 +32,18 @@ export function migrationStrategyForRouteControllers(
   /*
     Case 2: Passed the --pod flag to Ember CLI
   */
-  const filePaths2 = findFiles(
-    'tests/unit/controllers/**/controller-test.{js,ts}',
-    {
-      projectRoot,
-    },
-  );
+  const filePaths2 = findFiles('tests/unit/routes/**/route-test.{js,ts}', {
+    projectRoot,
+  });
 
   const filePathMap2 = filePaths2.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
-        directory: 'tests/unit/controllers',
-        file: 'controller-test',
+        directory: 'tests/unit/routes',
+        file: 'route-test',
       },
       replace: (key: string) => {
-        return `tests/unit/controllers/${key}-test`;
+        return `tests/unit/routes/${key}-test`;
       },
     });
